refactor(chat): extract artifact browser
